Changed to use float64 everywhere

+++ src/github.com/getlantern/flashlight/config/fetcher_test.go
@@ -15,7 +15,7 @@ func (uc *userConfig) GetToken() string {
 	return "token"
 }
 
-func (uc *userConfig) GetUserID() int {
+func (uc *userConfig) GetUserID() float64 {
 	return 10
 }
